Well the book was obviously wrong.

+++ drafts/0x03-cli/main.py
@@ -12,15 +12,12 @@ def cli():
 
 @click.command()
 @click.argument('file_path', type=click.Path(exists=True))
-@click.option('--all',
-              is_flag=True,
-              help='Show all metadata, including None values.')
+@click.option('--all', '-a', is_flag=True, help='Show all metadata.')
 @click.option('--existing',
+              '-e',
               is_flag=True,
-              help='Show existing metadata, excluding "art" and None values.')
-@click.option('--missing',
-              is_flag=True,
-              help='Show missing metadata, including None values.')
+              help='Show only existing metadata.')
+@click.option('--missing', '-m', is_flag=True, help='Show missing metadata.')
 def show(file_path, all, existing, missing):
     """Show metadata for a media file."""
     track = TrackInfo(file_path)
@@ -42,16 +39,39 @@ def show(file_path, all, existing, missing):
 def update(file_path, updates):
     """Update metadata for a media file."""
     track = TrackInfo(file_path)
-    track.update_metadata(updates)
-    track.save()
+    md_pre_update = track.as_dict()
+
+    track.batch_update_metadata(updates)
+
+    track.has_changes(track, md_pre_update)
+
+    if track.has_changes:
+        click.echo(f"Metadata changes for {track.metadata.filename}:")
+        for key, value in track.as_dict().items():
+            if key != "images" and key in md_pre_update and md_pre_update[
+                    key] != value:
+                click.echo(f"{key}: {md_pre_update[key]} -> {value}")
+
+        if click.confirm("Do you want to save these changes?"):
+            track.save()
+            click.echo("Changes saved.")
+        else:
+            click.echo("Changes not saved.")
+    else:
+        click.echo("No changes to save.")
 
 
 @click.command()
 @click.argument('file_path', type=click.Path(exists=True))
 def delete(file_path):
-    """Delete metadata for a media file."""
+    """Delete's all metadata from the media file."""
     track = TrackInfo(file_path)
-    track.delete()
+    if click.confirm(
+            f"Confirm removal of all {track.metadata.filename} tags?"):
+        track.delete()
+        click.echo("Metadata tags removed!.")
+    else:
+        click.echo("Deletion canceled.")
 
 
 def show_menu():
